Add files via upload
The goal of Employee Management System is help to attracting, developing, and retaining employees and also by managing them effectively. It aims to achieve a fit between managing the organization's employees and the overall strategic direction of the organization. EMS planning deals with recognizing and fulfilling the employee needs of an organization. In order to meet the effectiveness objective of management control, the Admin/HR plan should be formulated in such a way that it conforms to the organizational strategies and corporate plans. At the micro level, the organization needs to do career planning for employees and succession planning for key positions in the organization. At a macro level, the organization needs to carry out workforce (manpower) planning and salary planning. The nature of the HR planning process is primarily dependent upon the organizational size and the hierarchical structure within the organization.


Software Uses:

Operating System	:    Windows 10
Front-End		:    HTML, CSS
Back-End		        :    Python 3.10 , MYSQL 
Framework               :    Django 4.0